+++ erppeek.py
@@ -59,7 +59,7 @@ except ImportError:     # Python 2.5
         return _convert(node_or_string)
 
 
-__version__ = '1.4.4.dev0'
+__version__ = '1.4.4'
 __all__ = ['Client', 'Model', 'Record', 'RecordList', 'Service',
            'format_exception', 'read_config', 'start_openerp_services']